requirements.txt
#pip install -r requirements.txt

# GUI and file management
tkinterdnd2

# Image handling
pillow

# Version parsing
packaging

# Image Analysis
opencv-python

# Image Analysis
numpy